+++ test/mock-data/common-data-sets.js
@@ -9,7 +9,7 @@ exports.IntegerDataSet = [34, '65', 46677]
 exports.NotBooleanDataSet = ['ANC', '3453', 'xxx', '%%', ' ', 567, 3.3, [], {}, { x: 1 }, dummyFunction, class { }]
 exports.NotIntegerDataSet = ['ANC', ' ', '^^', true, 3.3, {}, [], { x: 1 }, dummyFunction, class { }]
 exports.NotFunctionDataSet = ['ANC', 3453, 3.6, true, [], {}]
-
+exports.NotObjectDataSet = ['ABC', 3, 3.3, [], dummyFunction, class { }]
 exports.CorrectPortNumber = [1, 65535, 22, 300, 6455]
 exports.IncorrectPortNumber = [0, 65536, 100100]